move the retrival logic to the controllers
