+++ src/app/modules/audio/types.ts
@@ -1,3 +1,6 @@
+// eslint-disable-next-line import/no-extraneous-dependencies
+import { ModuleEndpointContext } from 'lisk-framework';
+
 export interface Audio {
   name: string;
   releaseYear: string;
@@ -48,4 +51,9 @@ export interface SetAttributesCommandParams {
   artistName: string;
   genre: number[];
   audioID: Buffer;
+}
+
+export interface Store<Entity> {
+  get:  (context: ModuleEndpointContext, key: Buffer) => Promise<Entity>;
+  has: (context: ModuleEndpointContext, key: Buffer) => Promise<boolean>;
 }
